Move setting combination strategy into test
